+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/streaming/ClientStreamingQuerySuite.scala
@@ -28,7 +28,7 @@ import org.scalatest.concurrent.Futures.timeout
 import org.scalatest.time.SpanSugar._
 
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.{AnalysisException, DataFrame, Dataset, ForeachWriter, Row, SaveMode, SparkSession, SQLHelper}
+import org.apache.spark.sql.{DataFrame, ForeachWriter, Row, SparkSession, SQLHelper}
 import org.apache.spark.sql.connect.client.util.QueryTest
 import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.functions.window
@@ -351,26 +351,6 @@ class ClientStreamingQuerySuite extends QueryTest with SQLHelper with Logging {
 
       q.stop()
     }
-  }
-
-  test("deduplicate in batch DataFrame") {
-    def testAndVerify(df: Dataset[_]): Unit = {
-      val exc = intercept[AnalysisException] {
-        df.write.format("noop").mode(SaveMode.Append).save()
-      }
-
-      assert(exc.getMessage.contains("dropDuplicatesWithinWatermark is not supported"))
-      assert(exc.getMessage.contains("batch DataFrames/DataSets"))
-    }
-
-    val result = spark.range(10).dropDuplicatesWithinWatermark()
-    testAndVerify(result)
-
-    val result2 = spark
-      .range(10)
-      .withColumn("newcol", col("id"))
-      .dropDuplicatesWithinWatermark("newcol")
-    testAndVerify(result2)
   }
 }
